Reverts unneeded changes.
